DataTrigger

+++ Altkom.Shop.FakeServices/Fakers/ProductFaker.cs
@@ -19,7 +19,7 @@ namespace Altkom.Shop.FakeServices.Fakers
             RuleFor(p => p.Color, f => f.Commerce.Color());
             RuleFor(p => p.Name, f => f.Commerce.ProductName());
             RuleFor(p => p.UnitPrice, f => decimal.Parse(f.Commerce.Price()));
-            RuleFor(p => p.Weight, f => f.Random.Float(1, 100));
+            RuleFor(p => p.Weight, f => f.Random.Float(1, 1000));
         }
     }
 }
